Refactor binary API (remove Blob, add normal array).
There really isn't much point supporting Blob and ArrayBuffers since
browsers with Blob/ArrayBuffer support will almost certainly have native WebSocket
support.

However, supporting the ability to send binary frames is still really
useful so this adds a binaryType of "array" which is not in the API
spec but allows web-socket-js to be used to support older browsers.
The Javascript will just send an array of numbers (0-255) and this
will be converted to raw binary data.

I've retained ArrayBuffer support since this is not much different
than normal array support and it does allow the possibility where
a browser might be released with ArrayBuffers without WebSocket
(possible with IE :-( ). Also, it is useful in the case where
a developer wants to force web-socket-js even when the browser
supports WebSockets natively. For example, this would allow the new
HyBi protocol and binary interface to be used in browsers with the
Hixie and older API (Chrome 5-13, Firefox 4-5, Opera 11).

Also, getting rid of the Blob support allows refactoring to remove the
async send code and go back to synchronized send which is much
cleaner.

One final change is due to the fact that new WebSocket API spec
doesn't have an obvious way to detect the binary API without first
instantiating a WebSocket object. I've changed the init and close code
so that the async __flash.create() task is cancelled in close() so
that you can do something like this:

    var wstest = new WebSocket('ws://localhost:57111');
    if (typeof(wstest.binaryType) !== "undefined") {
        ...
    }
    wstest.close();
    wstest = null;

Without the change, a Flash error occurs later when the connect fails
(because the close() call doesn't cancel the connect).